add analyzer
